requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==4.0.0
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.1.0
backoff==2.2.1
backports.zoneinfo==0.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.6.2
blis==0.4.1
boilerpy3==1.0.6
boto3==1.28.62
botocore==1.31.62
CacheControl==0.13.1
cachetools==5.3.1
catalogue==1.0.2
cattrs==23.1.2
certifi==2023.7.22
cffi==1.16.0
chainlit==0.7.1
chardet==5.2.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.13
click==8.1.7
coloredlogs==15.0.1
construct==2.5.3
contourpy==1.1.1
cryptography==41.0.4
ctransformers==0.2.27
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.5.14
datasets==2.14.6
deep-translator==1.11.4
Deprecated==1.2.14
dill==0.3.7
dipy==1.7.0
docopt==0.6.2
emoji==2.8.0
entrypoints==0.4
Events==0.5
exceptiongroup==1.1.3
faiss-cpu==1.7.4
farm-haystack==1.21.2
fastapi==0.100.1
fastapi-socketio==0.0.10
filelock==3.12.4
filetype==1.2.0
firebase-admin==6.2.0
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
ftfy==5.9
gensim==4.3.2
gitdb==4.0.10
GitPython==3.1.37
google-api-core==2.12.0
google-api-python-client==2.104.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
graphlib-backport==1.0.3
greenlet==2.0.2
grpcio==1.59.0
grpcio-status==1.59.0
h11==0.14.0
h5py==3.10.0
haystack==0.42
httpcore==0.18.0
httplib2==0.22.0
httptools==0.6.0
httpx==0.25.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
inflect==7.0.0
install==1.3.5
InstructorEmbedding==1.0.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langchain==0.0.306
langdetect==1.0.9
langsmith==0.0.41
Lazify==0.4.0
lazy-imports==0.3.1
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.3
mdurl==0.1.2
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.1
nibabel==5.1.0
nltk==3.8.1
nodeenv==1.8.0
num2words==0.5.12
numexpr==2.8.6
numpy==1.24.4
onnxruntime==1.16.0
openai==0.27.6
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-instrumentation==0.40b0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pefile==2023.2.7
Pillow==10.0.1
pkgutil_resolve_name==1.3.10
plac==1.1.3
platformdirs==3.11.0
posthog==3.0.2
preshed==3.0.9
prisma==0.10.0
prompthub-py==4.0.0
proto-plus==1.22.3
protobuf==4.24.4
pulsar-client==3.3.0
py-cpuinfo==9.0.0
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.13
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
Pympler==1.0.1
pyparsing==3.1.1
PyPDF2==3.0.1
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.7.1
python-graphql-client==0.4.3
python-iso639==2023.6.15
python-magic==0.4.27
python-multipart==0.0.6
python-ptrace==0.9.8
python-socketio==5.9.0
pytokenizations==0.8.4
pytube==15.0.0
pytz==2023.3.post1
PyYAML==6.0.1
quantulum3==0.9.0
rank-bm25==0.2.2
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-cache==0.9.8
rich==13.6.0
rpds-py==0.10.3
rsa==4.9
s3transfer==0.7.0
sacremoses==0.0.53
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.10.1
seaborn==0.13.0
semver==3.0.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
simple-websocket==0.10.1
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
spacy==2.2.4
spacy-transformers==0.6.0
SQLAlchemy==2.0.21
srsly==1.0.7
sseclient-py==1.8.0
st-pages==0.4.5
starlette==0.27.0
streamlit==1.27.2
streamlit-chat==0.1.1
streamlit-player==0.1.5
sympy==1.12
syncer==2.0.3
tabulate==0.9.0
tenacity==8.2.3
textblob==0.17.1
thinc==7.4.0
threadpoolctl==3.2.0
tiktoken==0.5.1
TIME-python==0.0.17
tokenizers==0.14.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.0
torch==2.1.0
torchcontrib==0.0.2
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
transformers==4.34.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
unstructured==0.10.18
uptrace==1.20.0
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==1.26.17
uvicorn==0.23.2
uvloop==0.17.0
validators==0.22.0
wasabi==0.10.1
watchfiles==0.20.0
wcwidth==0.2.8
websockets==11.0.3
wrapt==1.15.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
